+++ src/Event/Cursor/PositionY.cpp
@@ -27,7 +27,7 @@
 Event::Cursor::PositionY::PositionY(std::string position) : AbstractField{"PositionY", std::move(position)} {
 }
 
-Event::Cursor::PositionY::PositionY(int position) : PositionY{std::to_string(position)} {
+Event::Cursor::PositionY::PositionY(double position) : PositionY{std::to_string(position)} {
 }
 
 Event::Cursor::PositionY::PositionY() : PositionY{""} {
